pyproject.toml
[tool.poetry]
name = "rl rubicks cube"
version = "0.1.0"
description = ""
authors = ["shreyan241 <shreyansood@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "1.24.4"
matplotlib = "^3.9.2"
flake8 = "^7.1.1"
tqdm = "^4.66.5"
flask = "^3.0.3"
google-cloud-storage = "^2.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
